refactor(e2e-suite): split off setup methods from suiteBaseFixture
